refactored editor
