+++ .gitignore
@@ -132,3 +132,4 @@ dmypy.json
 *.xml
 *.iml
 .idea
+.island
